Refactor store utilities
